DB: Connect and insert into database
Implemented connecting into SQL database and inserting records via the
form. The frontend is present in /views directory. views/index.html
handles the form which will be inserted into the database and
views/uploaded.ejs displays the information about the query that is
inserinserted. Database operations are handled in the /routes
directory.

The db_info_template.js contains the template of the database
information and the placeholder is needed to be replaced with the
appropriate values and the file should be renamed to db_info.js

Implemented stuffs:
* render the frontend present in views/
* Connect to the MySQL database
* Handle queries to insert the informations

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>